refactor: Move test blocks and toolbox to test/
AFAICT these are not supposed to be part of the actual plugin,
so it makes sense to move them to test/.
